+++ pom.xml
@@ -382,8 +382,6 @@
         <groovyVersion>4.0.15</groovyVersion>
         <seleniumVersion>4.15.0</seleniumVersion>
         <slf4jVersion>2.0.9</slf4jVersion>
-        <tilesVersion>3.0.8</tilesVersion>
-        <tilesRequestVersion>1.0.7</tilesRequestVersion>
         <antlrVersion>4.13.1</antlrVersion>
         <servletVersion>4.0.4</servletVersion>
         <jspVersion>2.3.6</jspVersion>
@@ -518,7 +516,7 @@
         <module>scripting</module>
         <module>taglib</module>
         <module>tiles</module>
-        <module>tiles2</module>
+        <module>tiles3</module>
     </modules>
 
     <licenses>
@@ -1157,32 +1155,6 @@
                 <version>${groovyVersion}</version>
             </dependency>
             <dependency>
-               <groupId>org.apache.tiles</groupId>
-               <artifactId>tiles-api</artifactId>
-               <version>${tilesVersion}</version>
-            </dependency>
-            <dependency>
-               <groupId>org.apache.tiles</groupId>
-               <artifactId>tiles-core</artifactId>
-               <version>${tilesVersion}</version>
-               <exclusions>
-                  <exclusion>
-                     <groupId>org.slf4j</groupId>
-                     <artifactId>jcl-over-slf4j</artifactId>
-                  </exclusion>
-               </exclusions>
-            </dependency>
-            <dependency>
-               <groupId>org.apache.tiles</groupId>
-               <artifactId>tiles-request-api</artifactId>
-               <version>${tilesRequestVersion}</version>
-            </dependency>
-            <dependency>
-               <groupId>org.apache.tiles</groupId>
-               <artifactId>tiles-request-servlet</artifactId>
-               <version>${tilesRequestVersion}</version>
-            </dependency>
-            <dependency>
                 <groupId>${project.groupId}</groupId>
                 <artifactId>struts-blank</artifactId>
                 <version>${project.version}</version>
@@ -1275,7 +1247,7 @@
             </dependency>
             <dependency>
                 <groupId>${project.groupId}</groupId>
-                <artifactId>struts-tiles2</artifactId>
+                <artifactId>struts-tiles3</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
