+++ tiles3/pom.xml
@@ -1,0 +1,140 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+    $Id$
+
+    Licensed to the Apache Software Foundation (ASF) under one or more
+    contributor license agreements.  See the NOTICE file distributed with
+    this work for additional information regarding copyright ownership.
+    The ASF licenses this file to You under the Apache License, Version 2.0
+    (the "License"); you may not use this file except in compliance with
+    the License.  You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+   <parent>
+      <groupId>io.github.weblegacy</groupId>
+      <artifactId>struts-parent</artifactId>
+      <version>1.4.6-SNAPSHOT</version>
+   </parent>
+
+   <modelVersion>4.0.0</modelVersion>
+   <artifactId>struts-tiles3</artifactId>
+   <packaging>jar</packaging>
+   <name>Struts Tiles 3 integration</name>
+
+   <properties>
+      <struts.osgi.symbolicName>org.apache.struts.tiles3</struts.osgi.symbolicName>
+      <tiles3Version>3.0.8</tiles3Version>
+      <tilesRequestVersion>1.0.7</tilesRequestVersion>
+   </properties>
+
+   <dependencies>
+      <dependency>
+         <groupId>${project.groupId}</groupId>
+         <artifactId>struts-core</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>${project.groupId}</groupId>
+         <artifactId>struts-core</artifactId>
+         <classifier>tests</classifier>
+         <type>test-jar</type>
+      </dependency>
+      <dependency>
+         <groupId>jakarta.servlet</groupId>
+         <artifactId>jakarta.servlet-api</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>jakarta.servlet.jsp</groupId>
+         <artifactId>jakarta.servlet.jsp-api</artifactId>
+         <scope>test</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.tiles</groupId>
+         <artifactId>tiles-api</artifactId>
+         <version>${tiles3Version}</version>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.tiles</groupId>
+         <artifactId>tiles-core</artifactId>
+         <version>${tiles3Version}</version>
+         <exclusions>
+            <exclusion>
+               <groupId>org.slf4j</groupId>
+               <artifactId>jcl-over-slf4j</artifactId>
+            </exclusion>
+         </exclusions>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.tiles</groupId>
+         <artifactId>tiles-request-api</artifactId>
+         <version>${tilesRequestVersion}</version>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.tiles</groupId>
+         <artifactId>tiles-request-servlet</artifactId>
+         <version>${tilesRequestVersion}</version>
+      </dependency>
+      <dependency>
+         <groupId>commons-beanutils</groupId>
+         <artifactId>commons-beanutils</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>io.github.weblegacy</groupId>
+         <artifactId>commons-chain</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>org.junit.jupiter</groupId>
+         <artifactId>junit-jupiter-api</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>org.slf4j</groupId>
+         <artifactId>slf4j-api</artifactId>
+      </dependency>
+      <!-- Needed for commons-beanutils -->
+      <dependency>
+         <groupId>org.slf4j</groupId>
+         <artifactId>jcl-over-slf4j</artifactId>
+      </dependency>
+   </dependencies>
+
+   <reporting>
+      <plugins>
+         <plugin>
+            <groupId>io.github.weblegacy</groupId>
+            <artifactId>taglib-maven-plugin</artifactId>
+            <configuration>
+               <srcDir>${project.basedir}/src/main/resources/META-INF/tld</srcDir>
+               <parseHtml>true</parseHtml>
+            </configuration>
+         </plugin>
+      </plugins>
+   </reporting>
+
+   <profiles>
+       <profile>
+        <id>pre-assembly</id>
+        <build>
+        <plugins>
+          <plugin>
+            <artifactId>maven-javadoc-plugin</artifactId>
+          </plugin>
+          <plugin>
+            <artifactId>maven-source-plugin</artifactId>
+          </plugin>
+       </plugins>
+       </build>
+      </profile>
+    </profiles>
+
+</project>
