+++ tiles3/src/main/resources/NOTICE.txt
